+++ shopify_auth/urls.py
@@ -5,6 +5,5 @@ from . import views
 urlpatterns = [
   path('finalize/',     views.finalize),
   path('authenticate/', views.authenticate),
-  path('check-cookie',  views.check_cookie),
   path('',              views.login),
 ]